Unifying the way we create faces with ccw normals
