+++ addon/algorithms/geodesic_edge_flipping.py
@@ -11,6 +11,7 @@ C++ implementation of the algorithm can be found here:
 https://github.com/nmwsharp/flip-geodesics-demo
 '''
 
+from ..utility.geometry import create_face_with_ccw_normal, get_angle_signed
 from enum import Enum
 from functools import reduce
 from math import fabs, inf, degrees, pi
@@ -18,7 +19,7 @@ from queue import PriorityQueue
 from typing import Tuple
 
 from bmesh.types import BMEdge, BMFace, BMVert, BMesh
-from mathutils import Matrix, Vector
+from mathutils import Vector
 from mathutils.geometry import intersect_line_line
 
 EPS_ANGLE: float = 1e-5
@@ -338,15 +339,6 @@ def flip_edge(bm: BMesh, e: BMEdge, pivot_vert: BMVert) \
         return (e2, e1, updated_edge)
 
 
-def create_face_with_ccw_normal(bm: BMesh,
-                                v1: BMVert, v2: BMVert, v3: BMVert) -> BMFace:
-    # Detect ccw
-    if get_angle_signed((v1.co-v2.co), (v3.co-v2.co), v2.normal) < 0:
-        return bm.faces.new((v3, v2, v1))
-    else:
-        return bm.faces.new((v1, v2, v3))
-
-
 def get_opposed_vert(face: BMFace, edge: BMEdge) -> BMVert:
     return [v for v in face.verts if v not in edge.verts][0]
 
@@ -439,22 +431,6 @@ def get_angles_signed(start_edge: BMEdge, end_edge: BMEdge) \
     # )
 
     return result
-
-
-def get_angle_signed(v1: Vector, v2: Vector, n: Vector) -> float:
-    # Implementing this idea:
-    # https://math.stackexchange.com/questions/1027476/calculating-clockwise-anti-clockwise-angles-from-a-point
-    matrix: Matrix = Matrix((v1, v2, n))
-    matrix.transpose()
-    det = matrix.determinant()
-    angle = v1.angle(v2)
-
-    if det < 0:  # clockwise
-        return angle
-    elif det > 0:  # anticlockwise
-        return -angle
-    else:
-        return 0
 
 
 def get_vectors(start_edge, end_edge):
